Refactor sources to keep C kernel small
After this commit, the C kernel only consists of

* src/rtosc.c
* src/dispatch.c
* src/rtosc-time.c (functions to handle OSC time format)

These files have been invalidly added to the C kernel in 2017
(60297243ae0b1600ef24379c77c5546d6822e333) and this is now fixed.
