+++ src/cpp/arg-val-math.c
