+++ src/cpp/ports-runtime.cpp
@@ -1,4 +1,4 @@
-#include "../util.h"
+#include "util.h"
 #include <cstdarg>
 #include <cstring>
 #include <cassert>
