+++ src/cpp/util.c
