+++ src/cpp/util.h
