+++ src/cpp/version.c.in
